module4-task1: refactor
