+++ src/mock/user.js
@@ -7,11 +7,4 @@ const getUserData = () => ({
   rating: `Movie Buff`
 });
 
-class User {
-  constructor(userData) {
-    this.avatar = userData.avatar;
-    this.rating = userData.rating;
-  }
-}
-
-export const user = new User(getUserData());
+export const user = getUserData();
